libtysila5: began x86_64 support libtysila5: added custom attribute support TableMap: converted default integer type to long
git-svn-id: https://www.tysos.org/svn/branches/tysila3@681 21779fe5-bacf-4c11-b3ef-82bd0e1f58e8

+++ TableMap/Program.cs
@@ -418,7 +418,7 @@ namespace TableMap
                         ToByteArray(a, ret);
                     break;
                 case Expression.EvalResult.ResultType.Int:
-                    CompressInt(v.intval, ret);
+                    CompressInt((int)v.intval, ret);
                     /*ret.Add((byte)(v.intval & 0xff));
                     ret.Add((byte)((v.intval >> 8) & 0xff));
                     ret.Add((byte)((v.intval >> 16) & 0xff));
